week 4 practices & projects

+++ week-2/project_1.rs
@@ -5,7 +5,7 @@ fn main() {
     let r: f64 = 10.0;
 
     // amount
-    let a = p * ((1.0 + (r / 100.0)) as f64).powf(t);
+    let a = p * (1.0 + (r / 100.0)).powf(t);
     println!("Amount: {}", a);
 
     // compound interest
